pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Agentic AI chatbot"
readme = "README.md"
authors = [
    { name = "Santosh Shrestha", email = "iam@santoshsth.com.np" }
]
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.42.0",
    "openai>=1.99.9",
    "pypdf>=6.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
]

[project.scripts]
chatbot = "chatbot:main"

[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"
